requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
bcrypt==3.2.0
PyJWT==2.1.0 